feat:完成会员添加功能

+++ src/main/java/com/lyr/busticketsystem/entity/Announcement.java
@@ -19,7 +19,7 @@ public class Announcement implements Serializable {
      * 公告ID（主键）
      */
     @TableId(value = "announcement_id", type = IdType.AUTO)
-    private Integer announcementId;
+    private Long announcementId;
 
     /**
      * 公告标题
